requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
libsql-client
requests
pycryptodome
aiohttp
